Fix: タスクメンバー選択機能の修正
